Fix XP threshold accumulation and rebalance node XP rewards
Ensures XP threshold accumulation uses integer values by casting the floor result to int, preventing incorrect accumulation due to floating-point types.

Rebalances experience rewards for several resource nodes (copper, beech, birch, oak) by lowering their experienceReward values to adjust progression pacing.

+++ plugins/skills/src/main/java/org/runetale/skills/service/XpService.java
@@ -53,7 +53,7 @@ public class XpService {
 		int[] thresholds = new int[MAX_LEVEL + 1];
 		int points = 0;
 		for (int level = 1; level <= MAX_LEVEL; level++) {
-			points += Math.floor((double) level + 300.0 * Math.pow(2.0, (double) level / 7.0));
+			points += (int) Math.floor((double) level + 300.0 * Math.pow(2.0, (double) level / 7.0));
 			if (level == 1) {
 				thresholds[level] = 0;
 			} else {
